🏗️ Reduce `compass.core` API surface area
