+++ compass.core/src/compass/core/__init__.py
@@ -2,116 +2,17 @@
 
 This module exposes the public api for CI core
 """
-from __future__ import annotations
-
-from compass.core import logger
 from compass.core.__version__ import __version__
 from compass.core.__version__ import __version_info__
-from compass.core.errors import CompassAuthenticationError
-from compass.core.errors import CompassError
-from compass.core.errors import CompassNetworkError
-from compass.core.errors import CompassPermissionError
-from compass.core.errors import CompassReportError
-from compass.core.errors import CompassReportPermissionError
-from compass.core.hierarchy import Hierarchy
-from compass.core.logon import Logon
-from compass.core.people import People
-from compass.core.reports import Reports
-from compass.core.schemas.hierarchy import DescendantData
-from compass.core.schemas.hierarchy import HierarchyLevel
-from compass.core.schemas.hierarchy import HierarchyMember
-from compass.core.schemas.hierarchy import HierarchySection
-from compass.core.schemas.hierarchy import HierarchyUnit
-from compass.core.schemas.hierarchy import HierarchyUnitMembers
-from compass.core.schemas.hierarchy import UnitData
-from compass.core.schemas.logon import CompassProps
-from compass.core.schemas.member import MemberAward
-from compass.core.schemas.member import MemberDetails
-from compass.core.schemas.member import MemberDisclosure
-from compass.core.schemas.member import MemberMandatoryTraining
-from compass.core.schemas.member import MemberPermit
-from compass.core.schemas.member import MemberRoleCore
-from compass.core.schemas.member import MemberRolePopup
-from compass.core.schemas.member import MemberRolesCollection
-from compass.core.schemas.member import MemberTrainingTab
-from compass.core.settings import Settings
-from compass.core.types.hierarchy import TYPES_HIERARCHY_LEVELS
-from compass.core.types.hierarchy import TYPES_UNIT_LEVELS
-from compass.core.types.logon import TYPES_ORG_LEVELS
-from compass.core.types.logon import TYPES_ROLE
-from compass.core.types.member import AddressData
-from compass.core.types.member import TYPES_ROLE_STATUS
-from compass.core.types.reports import TYPES_EXPORTED_REPORTS
-from compass.core.types.reports import TYPES_FORMAT_CODE
-from compass.core.types.reports import TYPES_FORMAT_CODES
-from compass.core.types.reports import TYPES_REPORTS
 
-
-class CompassInterface:
-    def __init__(self, user_props: Logon, /):
-        """This is the main (programmatic) interface to CI core."""
-        self.user = user_props
-        self.people = People(user_props)
-        self.hierarchy = Hierarchy(user_props)
-        self.reports = Reports(user_props)
-
-
-def login(username: str, password: str, /, *, role: str | None = None, location: str | None = None) -> CompassInterface:
-    """Log in to compass, return a compass.logon.Logon object.
-
-    This function is provided as a convenient interface to the logon module.
-    """
-    return CompassInterface(Logon.from_logon((username, password), role, location))
-
+from compass.core.interface import login
+from compass.core.logger import enable_debug_logging
 
 __all__ = (
     # public metadata
     "__version__",
     "__version_info__",
-    # public sub-modules
-    "logger",
-    # public classes
-    "CompassInterface",
-    "Logon",
-    "Hierarchy",
-    "People",
-    "Reports",
-    "Settings",
-    # public exceptions
-    "CompassError",
-    "CompassAuthenticationError",
-    "CompassNetworkError",
-    "CompassPermissionError",
-    "CompassReportError",
-    "CompassReportPermissionError",
     # public functions
+    "enable_debug_logging",
     "login",
-    # public types
-    "TYPES_EXPORTED_REPORTS",
-    "TYPES_FORMAT_CODE",
-    "TYPES_FORMAT_CODES",
-    "TYPES_HIERARCHY_LEVELS",
-    "TYPES_ORG_LEVELS",
-    "TYPES_REPORTS",
-    "TYPES_ROLE",
-    "TYPES_ROLE_STATUS",
-    "TYPES_UNIT_LEVELS",
-    "AddressData",
-    "CompassProps",
-    "DescendantData",
-    "HierarchyLevel",
-    "HierarchyMember",
-    "HierarchySection",
-    "HierarchyUnit",
-    "HierarchyUnitMembers",
-    "MemberAward",
-    "MemberDetails",
-    "MemberDisclosure",
-    "MemberMandatoryTraining",
-    "MemberPermit",
-    "MemberRoleCore",
-    "MemberRolePopup",
-    "MemberRolesCollection",
-    "MemberTrainingTab",
-    "UnitData",
 )
