+++ tests/test_compass_core_logger.py
@@ -2,7 +2,8 @@ import logging
 import pathlib
 import sys
 
-from compass.core import logger
+from compass.core.logger import logger
+from compass.core.logger import enable_debug_logging
 
 
 class TestLogger:
@@ -11,11 +12,11 @@ class TestLogger:
         log_file = tmp_path / "test_log.log"
 
         # When
-        logger.enable_debug_logging(log_file=log_file)
+        enable_debug_logging(log_file=log_file)
 
         # Then
-        assert logger.logger.level == logging.DEBUG
-        assert isinstance(logger.logger.handlers[0], logging.StreamHandler)
-        assert isinstance(logger.logger.handlers[1], logging.FileHandler)
-        assert logger.logger.handlers[0].stream == sys.stdout
-        assert "test_log.log" in logger.logger.handlers[1].baseFilename
+        assert logger.level == logging.DEBUG
+        assert isinstance(logger.handlers[0], logging.StreamHandler)
+        assert isinstance(logger.handlers[1], logging.FileHandler)
+        assert logger.handlers[0].stream == sys.stdout
+        assert "test_log.log" in logger.handlers[1].baseFilename
